// src/constants/enums.ts
export enum GameType {
  TreasureBox = '3-treasure box',
  BlindBox = '3-blind box',
  ThreeDTreausreBox = '3d treasure box',
  ScratchTheCard = 'scratch the card',
  Gachapon = 'gachapon'
}

export enum PrizeType {
  Coupon = 'coupon',
  FreeProduct = 'free product',
  CustomPrize = 'custom prize',
  NoPrize = 'no prize'
}